llm_backend/requirements.txt
fastapi
uvicorn
jinja2
requests
python-dotenv
openai
fastmcp
sqlalchemy
psycopg2-binary
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart